refactor data size, download speed and upload speed to double

+++ src/main/java/com/gigacapstone/billingservice/dto/BundlePackageDTO.java
@@ -13,7 +13,7 @@ public class BundlePackageDTO {
     int vatPercentage;
     ExpirationRate expirationRate;
     CallTime callTime;
-    int downloadSpeed;
-    int uploadSpeed;
-    int dataSize;
+    double downloadSpeed;
+    double uploadSpeed;
+    double dataSize;
 }
